Fix long long is 64 bit, not long.
- apparently int and long have the same size. Go figure.

+++ include/std/types.h
@@ -4,9 +4,9 @@
 typedef enum bool { true = 1, false = 0 } bool;
 
 typedef int int32;
-typedef long int64;
+typedef long long int64;
 
 typedef unsigned int uint32;
-typedef unsigned long uint64;
+typedef unsigned long long uint64;
 
 #endif /* STD_TYPES_H */
